+++ src/components/Checkbox.js
@@ -7,7 +7,7 @@ const Wrapper = styled.View`
   width: 24;
   height: 24;
   border-radius: 12;
-  background-color: ${({ color }) => color};
+  background-color: ${({ value, color, disabledColor }) => (value ? color : disabledColor)};
   border-width: 1;
   border-color: rgba(74, 74, 74, 0.2);
 `;
@@ -22,21 +22,25 @@ const Checkmark = styled(Ionicons).attrs({
   margin-left: 3;
 `;
 
-const Checkbox = ({ value, onPress, color }) => (
-  <Wrapper onPress={onPress} color={color}>
-    <Checkmark value={value} disabledColor={color} />
+const Checkbox = ({ value, onPress, color, disabledColor, disabledCheckmarkColor }) => (
+  <Wrapper onPress={onPress} color={color} disabledColor={disabledColor} value={value}>
+    <Checkmark value={value} disabledColor={disabledCheckmarkColor} />
   </Wrapper>
 );
 
 Checkbox.propTypes = {
   value: PropTypes.bool,
   color: PropTypes.string,
+  disabledColor: PropTypes.string,
+  disabledCheckmarkColor: PropTypes.string,
   onPress: PropTypes.func,
 };
 
 Checkbox.defaultProps = {
   value: false,
-  color: '#4494d3',
+  color: '#4494d3', // #4494d3
+  disabledColor: '#fff',
+  disabledCheckmarkColor: '#fff',
   onPress: undefined,
 };