+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - [FAQ] New FAQ page
 - [BurgerMenu] New Texts for all Menus
 
+### Changed
+- [Filter] changed colors of disabled Sub-Checkboxes [#533](https://github.com/demokratie-live/democracy-client/issues/533)
+
 ## [0.10.1] - 2018-09-17
 ### Added
 - Beta End Nag Screen